@@ -2,7 +2,7 @@
   "name": "lilt/craft-lilt-plugin",
   "description": "The Lilt plugin makes it easy for you to send content to Lilt for translation right from within Craft CMS.",
   "type": "craft-plugin",
-  "version": "3.6.0",
+  "version": "4.2.0",
   "keywords": [
     "craft",
     "cms",
@@ -22,10 +22,10 @@
     }
   ],
   "require": {
-    "craftcms/cms": "^3.7.0",
-    "php": ">=7.2.5",
+    "craftcms/cms": "^4.0.0",
     "lilt/lilt-connector-sdk-php": "0.1.0",
-    "ext-json": "*"
+    "ext-json": "*",
+    "php": "^8.0"
   },
   "require-dev": {
     "codeception/codeception": "^3.0 | ^4.0",
@@ -33,11 +33,10 @@
     "squizlabs/php_codesniffer": "^3.6",
     "codeception/module-yii2": "^1.1",
     "codeception/module-asserts": "^1.3",
-    "craftcms/redactor": "^2.10",
-    "fruitstudios/linkit": "1.1.12.1",
-    "percipioglobal/craft-colour-swatches": "^1.5",
-    "spicyweb/craft-neo": "^2.0",
-    "verbb/super-table": "^2.7",
+    "craftcms/redactor": "^2.0 || ^3.0",
+    "percipioglobal/craft-colour-swatches": "^1.0 || ^4.0",
+    "spicyweb/craft-neo": "3.1.5",
+    "verbb/super-table": "^3.0",
     "wiremock-php/wiremock-php": "^2.33"
   },
   "autoload": {
