@@ -1,11 +1,12 @@
 {
   "require": {
     "php": ">=7.0.0",
-    "craftcms/cms": "3.7.68",
-    "vlucas/phpdotenv": "^2.4.0",
-    "craftcms/redactor": "^2.8.7",
-    "verbb/super-table": "2.7.5.1",
-    "spicyweb/craft-neo": "^2.13"
+    "craftcms/cms": "^4.0.0",
+    "vlucas/phpdotenv": "2.6.7",
+    "craftcms/redactor": "*",
+    "verbb/super-table": "*",
+    "spicyweb/craft-neo": "*",
+    "yiisoft/yii2": "2.0.45"
   },
   "autoload": {
     "psr-4": {
@@ -27,7 +28,7 @@
   "config": {
     "optimize-autoloader": true,
     "platform": {
-      "php": "7.4"
+      "php": "8.0.26"
     },
     "allow-plugins": {
       "craftcms/plugin-installer": true,
