@@ -32,19 +32,19 @@ describe(
                 'Ready for review',
             );
 
-        cy.get('#translations-list th[data-title="Title"] div.element').
+        cy.get('#translations-list th[data-title="Translation"] div.element').
             invoke('attr', 'data-type').
             should('equal', 'lilthq\\craftliltplugin\\elements\\Translation');
 
-        cy.get('#translations-list th[data-title="Title"] div.element').
+        cy.get('#translations-list th[data-title="Translation"] div.element').
             invoke('attr', 'data-status').
             should('equal', 'ready-for-review');
 
-        cy.get('#translations-list th[data-title="Title"] div.element').
+        cy.get('#translations-list th[data-title="Translation"] div.element').
             invoke('attr', 'data-label').
             should('equal', 'The Future of Augmented Reality');
 
-        cy.get('#translations-list th[data-title="Title"] div.element').
+        cy.get('#translations-list th[data-title="Translation"] div.element').
             invoke('attr', 'title').
             should('equal',
                 'The Future of Augmented Reality – Happy Lager (en)');
@@ -68,7 +68,7 @@ describe(
             then((createdJobId) => {
               const appUrl = Cypress.env('APP_URL');
               cy.url().should(
-                  'equal',
+                  'contain',
                   `${appUrl}/admin/craft-lilt-plugin/job/edit/${createdJobId}`,
               );
             });
