@@ -8,6 +8,13 @@ describe(
       it('with copy slug disabled & enable after publish disabled', () => {
         const {jobTitle, slug} = generateJobData();
 
+        // cy.assertEntryContent(
+        //     ['en'],
+        //     'copy_source_text',
+        //     24,
+        // );
+        //
+        // return;
         cy.copySourceTextFlow({
           slug,
           entryLabel,
@@ -15,49 +22,7 @@ describe(
           copySlug: false,
           enableAfterPublish: false,
           languages: ['de', 'es', 'uk'],
-          batchPublishing: true
-        })
-      });
-
-      it('with copy slug disabled & enable after publish enabled', () => {
-        const {jobTitle, slug} = generateJobData();
-
-        cy.copySourceTextFlow({
-          slug,
-          entryLabel,
-          jobTitle,
-          copySlug: false,
-          enableAfterPublish: true,
-          languages: ['de', 'es', 'uk'],
-          batchPublishing: true
-        })
-      });
-
-      it('with copy slug enabled & enable after publish disabled', () => {
-        const {jobTitle, slug} = generateJobData();
-
-        cy.copySourceTextFlow({
-          slug,
-          entryLabel,
-          jobTitle,
-          copySlug: true,
-          enableAfterPublish: false,
-          languages: ['de', 'es', 'uk'],
-          batchPublishing: true
-        })
-      });
-
-      it('with copy slug enabled & enable after publish enabled', () => {
-        const {jobTitle, slug} = generateJobData();
-
-        cy.copySourceTextFlow({
-          slug,
-          entryLabel,
-          jobTitle,
-          copySlug: true,
-          enableAfterPublish: true,
-          languages: ['de', 'es', 'uk'],
-          batchPublishing: true
-        })
+          batchPublishing: true,
+        });
       });
     });
