@@ -7,13 +7,19 @@ namespace lilthq\craftliltplugin\services\handlers\field\copier;
 use Craft;
 use craft\base\ElementInterface;
 use craft\base\FieldInterface;
+use craft\elements\db\MatrixBlockQuery;
 use craft\elements\MatrixBlock;
 use craft\errors\InvalidFieldException;
+use craft\fields\Matrix;
 use lilthq\craftliltplugin\parameters\CraftliltpluginParameters;
 
 class MatrixFieldCopier implements FieldCopierInterface
 {
     /**
+     * @param FieldInterface|Matrix $field
+     * @param ElementInterface $from
+     * @param ElementInterface $to
+     * @return bool
      * @throws InvalidFieldException
      * @throws \Throwable
      */
@@ -27,6 +33,33 @@ class MatrixFieldCopier implements FieldCopierInterface
             return false;
         }
 
+        $this->removeBlocks($to, $field);
+
+        $serializedValue = $field->serializeValue($from->getFieldValue($field->handle), $from);
+
+        $prepared = [];
+        $i = 1;
+        foreach ($serializedValue as $item) {
+            $prepared[sprintf('new%d', $i++)] = $item;
+        }
+
+        $to->setFieldValues([$field->handle => $prepared]);
+
+        return true;
+    }
+
+    /**
+     * @param ElementInterface $to
+     * @param FieldInterface|Matrix $field
+     * @return void
+     * @throws InvalidFieldException
+     * @throws \Throwable
+     */
+    private function removeBlocks(ElementInterface $to, FieldInterface $field): void
+    {
+        /**
+         * @var MatrixBlockQuery $blocksQuery
+         */
         $blocksQuery = $to->getFieldValue($field->handle);
 
         /**
@@ -34,15 +67,14 @@ class MatrixFieldCopier implements FieldCopierInterface
          */
         $blocks = $blocksQuery->all();
 
-        Craft::$app->matrix->duplicateBlocks($field, $from, $to, false, false);
-        Craft::$app->matrix->saveField($field, $to);
-
         foreach ($blocks as $block) {
-            if ($block instanceof MatrixBlock) {
-                Craft::$app->getElements()->deleteElement($block, true);
+            if (!$block instanceof MatrixBlock) {
+                continue;
             }
+
+            Craft::$app->getElements()->deleteElement($block, true);
         }
 
-        return true;
+        Craft::$app->matrix->saveField($field, $to);
     }
 }
