@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace lilthq\craftliltplugin\services\handlers\field\copier;
 
+use benf\neo\elements\Block;
+use benf\neo\elements\db\BlockQuery;
+use benf\neo\Field;
 use Craft;
 use craft\base\ElementInterface;
 use craft\base\FieldInterface;
@@ -37,6 +40,48 @@ class NeoFieldCopier implements FieldCopierInterface
             return false;
         }
 
+        $this->removeBlocks($to, $field);
+
+        $serializedValue = $field->serializeValue($from->getFieldValue($field->handle), $from);
+
+        $prepared = [];
+        $i = 1;
+        foreach ($serializedValue as $item) {
+            $prepared[sprintf('new%d', $i++)] = $item;
+        }
+
+        $to->setFieldValues([$field->handle => $prepared]);
+
+        return true;
+    }
+
+    /**
+     * @param ElementInterface $to
+     * @param FieldInterface|Field $field
+     * @return void
+     * @throws InvalidFieldException
+     * @throws \Throwable
+     */
+    private function removeBlocks(ElementInterface $to, FieldInterface $field): void
+    {
+        /**
+         * @var BlockQuery $blocksQuery
+         */
+        $blocksQuery = $to->getFieldValue($field->handle);
+
+        /**
+         * @var Block[] $blocks
+         */
+        $blocks = $blocksQuery->all();
+
+        foreach ($blocks as $block) {
+            if (!$block instanceof Block) {
+                continue;
+            }
+
+            Craft::$app->getElements()->deleteElement($block, true);
+        }
+
         // Get the Neo plugin instance
         /** @var \benf\neo\Plugin $neoPluginInstance */
         $neoPluginInstance = call_user_func(['benf\neo\Plugin', 'getInstance']);
@@ -45,16 +90,7 @@ class NeoFieldCopier implements FieldCopierInterface
         /** @var \benf\neo\services\Fields $neoPluginFieldsService  */
         $neoPluginFieldsService = $neoPluginInstance->get('fields');
 
-        // Clear current neo field value
-        $neoField = $to->getFieldValue($field->handle);
-        foreach ($neoField as $block) {
-            Craft::$app->getElements()->deleteElement($block);
-        }
-        Craft::$app->getElements()->saveElement($to);
-
-        // Duplicate the blocks for the field
-        $neoPluginFieldsService->duplicateBlocks($field, $from, $to);
-
-        return true;
+        //Save field value
+        $neoPluginFieldsService->saveValue($field, $to);
     }
 }
